+++ src/main/php/lang/types/Byte.class.php
@@ -4,6 +4,8 @@
  * The Byte class wraps a value of the type byte 
  * 
  * Range: -2^7 - (2^7)- 1
+ *
+ * @deprecated Wrapper types will move to their own library
  */
 class Byte extends Number {